rename pages to index.tsx
