+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { Page404 } from './Page404/Page404';
-export { MainPage } from './MainPage/MainPage';
-export { AuthPage } from './AuthPage/AuthPage';
-export { InnerPage } from './InnerPage/InnerPage';
+export { Page404 } from './Page404';
+export { MainPage } from './MainPage';
+export { AuthPage } from './AuthPage';
+export { InnerPage } from './InnerPage';
